modified:   hoa.library/ext_win_debug/CL.read.1.tlog 	modified:   hoa.library/ext_win_debug/CL.write.1.tlog 	modified:   hoa.library/ext_win_debug/cl.command.1.tlog 	modified:   hoa.library/ext_win_debug/hoa.decoder~.idb 	modified:   hoa.library/ext_win_debug/hoa.decoder~.log 	modified:   hoa.library/ext_win_debug/link.command.1.tlog 	modified:   hoa.library/ext_win_debug/link.read.1.tlog 	modified:   hoa.library/ext_win_debug/link.write.1.tlog 	modified:   hoa.library/ext_win_release/CL.read.1.tlog 	modified:   hoa.library/ext_win_release/CL.write.1.tlog 	modified:   hoa.library/ext_win_release/cl.command.1.tlog 	modified:   hoa.library/ext_win_release/hoa.dac~.exp 	modified:   hoa.library/ext_win_release/hoa.dac~.lib 	modified:   hoa.library/ext_win_release/hoa.dac~.log 	modified:   hoa.library/ext_win_release/hoa.dac~.write.1.tlog 	modified:   hoa.library/ext_win_release/hoa.decoder~.log 	deleted:    hoa.library/ext_win_release/hoa.decoder~.unsuccessfulbuild 	modified:   hoa.library/ext_win_release/hoa.decoder~.write.1.tlog 	modified:   hoa.library/ext_win_release/link.command.1.tlog 	modified:   hoa.library/ext_win_release/link.read.1.tlog 	modified:   hoa.library/ext_win_release/link.write.1.tlog 	modified:   hoa.library/hoaBinaural/AmbisonicBinaural.cpp 	deleted:    hoa.library/hoaBinaural/AmbisonicBinaural.h 	modified:   hoa.library/hoaBinaural/ambisonicDecode.cpp 	deleted:    hoa.library/hoaBinaural/ambisonicDecode.h 	deleted:    hoa.library/hoaBinaural/fileLoader.c 	deleted:    hoa.library/hoaBinaural/fileLoader.h 	deleted:    hoa.library/hoaBinaural/gslMatrix.c 	deleted:    hoa.library/hoaBinaural/gslMatrix.h 	deleted:    hoa.library/hoaBinaural/hoa.binaural~.c 	modified:   hoa.library/hoaDac/hoa.dac~.c 	modified:   hoa.library/hoaDac/hoa.dac~.vcxproj 	modified:   hoa.library/hoaDecoder/ambisonicDecode.cpp 	modified:   hoa.library/hoaDecoder/ambisonicDecode.hpp 	modified:   hoa.library/hoaDecoder/hoa.decoder~.cpp 	modified:   hoa.library/hoaDecoder/hoa.decoder~.vcxproj 	deleted:    hoa.library/hoaDecoder/matrixHandling.h 	deleted:    hoa.library/hoaDecoder/matrixInverse.h 	deleted:    hoa.library/hoaEncoder/AmbisonicEncode.h 	deleted:    hoa.library/hoaEncoder/hoa.encoder~.c 	deleted:    hoa.library/hoaPlug/hoa.plug~.h 	modified:   hoa.library/hoaProjector/ambisonicProjector.cpp 	modified:   hoa.library/hoaProjector/ambisonicProjector.h 	deleted:    hoa.library/hoaProjector/gslMatrix.c 	deleted:    hoa.library/hoaProjector/gslMatrix.h 	modified:   hoa.library/hoaProjector/hoa.projector~.cpp 	modified:   hoa.library/hoaRecomposer/ambisonicRecomposer.cpp 	deleted:    hoa.library/hoaRecomposer/ambisonicRecomposer.h 	deleted:    hoa.library/hoaRecomposer/gslMatrix.c 	deleted:    hoa.library/hoaRecomposer/gslMatrix.h 	modified:   hoa.library/hoaRecomposer/hoa.recomposer~.cpp 	deleted:    hoa.library/hoaRecomposer/matrixHandling.h 	deleted:    hoa.library/hoaRecomposer/matrixInverse.h 	deleted:    hoa.library/hoaRotate/ambisonicRotate.h 	deleted:    hoa.library/hoaRotate/hoa.rotate~.c 	deleted:    hoa.library/hoaRotate/readme 	deleted:    sdk-build/hoa.*~.maxpat 	modified:   sdk-build/hoa.binaural~.maxhelp 	modified:   sdk-build/hoa.scope.maxhelp 	deleted:    sdk-build/hoaLibrary/operators/hoa.*~.maxpat

+++ hoa.library/hoaBinaural/ambisonicDecode.cpp
@@ -7,7 +7,7 @@
  *
  */
 
-#include "ambisonicDecode.h"
+#include "ambisonicDecode.hpp"
 
 ambisonicDecode::ambisonicDecode(int channels,int order): m_order(order), m_harmonics(order*2+1), m_channels(channels), m_optimId("basic")
 {
@@ -20,7 +20,7 @@ ambisonicDecode::ambisonicDecode(int channels,int order): m_order(order), m_harm
 	m_harmonicsIndex[0] = 0;
 	for(int i = 1; i < m_harmonics; i++)
 	{
-		m_harmonicsIndex[i] = (int)floor((i-1)/2) + 1;
+		m_harmonicsIndex[i] = (int)floor((float)(i-1) / 2.) + 1;
 		if (i % 2 == 1) 
 			m_harmonicsIndex[i] = - m_harmonicsIndex[i];
 	}
